chore(web): remove the v1 brain contract and the usage route
The hosted tier's first brain contract had no caller left: the desktop
reads the capabilities and speaks the second contract alone, and there is
no fallback to fall back to. The usage route had none either — nothing
reads today's counters over the wire.

Removing the first contract removes the turn authority with it: it existed
only to fix a toolset the service derived, and the desktop has named its
own tools since the second contract landed. `HOSTED_BRAIN_DEFAULTS` moves
to `brain-v2.ts`, the one file that still reads it, and the prose that
called the surviving contract "the second" now calls it the contract.
