+++ apps/web/server/README.md
@@ -185,37 +185,44 @@ backstop and should be configured with it.
 
 # Hosted brain inference
 
-`api/brain/respond.ts` runs one inference of Luke's brain on the deployment's
-own OpenAI key for a signed-in client that carries none of its own. It is an
-exact-path file like the voice route, resolved to a user through the same
-bearer seam, and it is the one hosted route with a raised function duration:
-`vercel.json` gives it 120 seconds so the 90-second
-upstream ceiling the brain shares with its keyed client can pass, and that
-`functions` entry names this route alone.
+`api/brain/capabilities.ts` and the four routes under `api/brain/v2/` run
+Luke's brain on the deployment's own OpenAI key for a signed-in client that
+carries none of its own. They are exact-path files like the voice route,
+resolved to a user through the same bearer seam, and the four POST routes are
+the only hosted routes with a raised function duration: `vercel.json` gives
+the inference, the compaction, and the token count 120 seconds so the
+90-second upstream ceiling the brain shares with its keyed client can pass,
+and the embedding 60.
 
-One HTTP request is one model call and nothing more. The client sends
-`{ authority, input }`: who opened the turn, and the Responses API input array
-it would have posted itself. `hostedBrainRequestFromWire` in `@sidecar/hosted`
-is the whole admission policy — the body is read as it streams and cut at 2 MiB
-whatever its `Content-Length` says, the array is capped at 2,000 items, each
-item must take one of the forms the brain replays, and a body carrying any
-other field (a model, instructions, tools, a store flag) is refused. The
-service then fixes the model, the instructions, the toolset for that
-authority, the output budget, and `store: false` from its own build and posts
-once. The answer is handed down as it came, once it is known to be a Responses
-payload every item of which the same admission would replay next turn; an
-answer this route could not replay is a 502, because the client would keep it
-verbatim and every later turn of that memory would fail here. The service runs
+One HTTP request is one model call and nothing more. A client GETs the
+capabilities first — the model, the operations, the registered tool names, the
+bounds, the reasoning efforts — and fails on a service that lacks them rather
+than falling back. It then POSTs one operation with the prompt it prepared and
+the tool names it means to offer. The readers in `brain-contract.ts` of
+`@sidecar/hosted` are the whole admission policy — the body is read as it
+streams and cut at 2 MiB whatever its `Content-Length` says, the prompt is
+bounded to its own 200,000-character envelope and refused past it rather than
+cut, the input array is capped at 2,000 items each of which must take one of
+the forms the brain replays, and every tool name must be one this service
+registers a schema for, so a caller can never upload a schema. The service
+then fixes the model, the upstream, its credential, the output budget's
+ceiling, and `store: false` from its own build and posts once. An inference's
+answer is handed down as it came, once it is known to be a Responses payload
+every item of which the same admission would replay next turn; an answer this
+route could not replay is a 502, because the client would keep it verbatim and
+every later turn of that memory would fail here. An explicit compaction's
+whole window is answered for the client to adopt as it came. The service runs
 no tool, holds no memory, reaches no provider, and stores and logs nothing of
 the request, the reply, or the encrypted compaction that travels in them.
 
 Each request spends the attention review meter before the upstream call, so a
-refused upstream still counts. `LUKE_BRAIN_MODEL` optionally overrides the
-model, under the name the desktop's keyed client honours; a blank value is
-treated as absent, and nothing in a request can name one. Without
-`OPENAI_API_KEY` the route answers 503 like the rest of the hosted tier. The
-existing mint and device routes are untouched by this route and keep their
-contracts for released clients.
+refused upstream still counts; an upstream that rate limits answers a bounded
+`Retry-After` the client cools down for, and the allowance was still spent.
+`LUKE_BRAIN_MODEL` optionally overrides the model, under the name the
+desktop's keyed client honours; a blank value is treated as absent, and
+nothing in a request can name one. Without `OPENAI_API_KEY` the routes answer
+503 like the rest of the hosted tier. The existing mint and device routes are
+untouched by these routes and keep their contracts for released clients.
 
 # Provider key vault
