+++ apps/web/vercel.json
@@ -2,7 +2,6 @@
   "$schema": "https://openapi.vercel.sh/vercel.json",
   "buildCommand": "pnpm db:migrate && pnpm auth:seed && pnpm build",
   "functions": {
-    "api/brain/respond.ts": { "maxDuration": 120 },
     "api/brain/v2/respond.ts": { "maxDuration": 120 },
     "api/brain/v2/compact.ts": { "maxDuration": 120 },
     "api/brain/v2/count-tokens.ts": { "maxDuration": 120 },
