+++ packages/brain/src/instructions.ts
@@ -1,5 +1,3 @@
-import { LUKE_PERSONA } from "@sidecar/guide";
-import { PROMPT_SAFETY_LINES } from "@sidecar/runtime";
 import { BRAIN_INPUT_MARKER } from "./input-items.js";
 import { BRAIN_TOOL, maximumBriefingLength } from "./tools.js";
 
@@ -9,9 +7,7 @@ import { BRAIN_TOOL, maximumBriefingLength } from "./tools.js";
  * because it names the fixed vocabulary — the input markers, the tool names,
  * the briefing bound — that the code fixes. The persona is SOUL.md's when a
  * workspace stands, and the prompt builder takes `brainToolNotes` as its
- * tool-notes section beside the workspace files and its own safety section;
- * `brainInstructions` joins persona, notes, and safety for the first hosted
- * contract, which composes no workspace prompt.
+ * tool-notes section beside the workspace files and its own safety section.
  */
 
 const ROLE_LINES: readonly string[] = [
@@ -109,9 +105,4 @@ const TOOL_LINES: readonly string[] = [
 /** The build's lines about the role, the turns, and the tools, for the prompt builder's tool-notes section. */
 export function brainToolNotes(): readonly string[] {
   return [...ROLE_LINES, "", ...TURN_LINES, "", ...TOOL_LINES];
-}
-
-/** The standing instructions of the first hosted contract: the shared persona, the build's own lines, the safety lines. */
-export function brainInstructions(): string {
-  return [LUKE_PERSONA, "", ...brainToolNotes(), "", ...PROMPT_SAFETY_LINES].join("\n");
 }
