+++ packages/hosted/src/index.ts
@@ -28,9 +28,6 @@ export {
   hostedBrainRespondRequestFromWire,
 } from "./brain-contract.js";
 export {
-  BRAIN_TURN_AUTHORITY,
-  type BrainTurnAuthority,
-  brainTurnAuthorityFromWire,
   DEVICE_PLATFORM,
   DEVICE_TOKEN_BOUNDS,
   type DevicePlatform,
@@ -48,24 +45,20 @@ export {
   type HostedActResult,
   type HostedActWorkspaceAnswer,
   type HostedApiError,
-  type HostedBrainRequest,
   type HostedConversationAnswer,
   type HostedConversationMessage,
   type HostedMintAnswer,
   type HostedProjectsAnswer,
   type HostedQuota,
-  type HostedUsageAnswer,
   type HostedWorkspaceAgentModels,
   type HostedWorkspaceProject,
   hostedActAnswerFromWire,
   hostedActWorkspaceAnswerFromWire,
-  hostedBrainRequestFromWire,
   hostedConversationAnswerFromWire,
   hostedErrorFromWire,
   hostedMintAnswerFromWire,
   hostedProjectsAnswerFromWire,
   hostedQuotaFromWire,
-  hostedUsageAnswerFromWire,
   isDevicePlatform,
   isPushEnvironment,
   isVaultProviderId,
